added the visable wikis logic to the wikis model

+++ app/controllers/colaborators_controller.rb
@@ -1,9 +1,12 @@
 # Public: This controller handles adding and removing colaborators for wikis
 class ColaboratorsController < ApplicationController
+  after_action :verify_authorized
+
   def create
     user = User.find_by_email(params[:colaborator_email])
     wiki = Wiki.find(params[:wiki_id])
     colaborator = Colaborator.new(user: user, wiki: wiki)
+    authorize colaborator
 
     if colaborator.save
       flash[:notice] = params[:colaborator_email] +
@@ -17,8 +20,9 @@ class ColaboratorsController < ApplicationController
   end
 
   def destroy
-    colaborator = Colaborator.find(params[:id])
     wiki = Wiki.find(params[:wiki_id])
+    colaborator = wiki.colaborators.find(params[:id])
+    authorize colaborator
 
     if colaborator.destroy
       flash[:notice] = 'Colaborator was deleted successfully'
